Took all day to create withdrawal from the inventory, when a sell is made. Now it checks over the lots and take the quantity, and cost from the lot to calculate the Profit of the sale. this will help to decide wich lot to sell.

+++ WebSalesMvcWithAngular/Models/Department.cs
@@ -34,7 +34,7 @@ namespace WebSalesMvc.Models
             Sellers.Add(seller);
         }
 
-        public double TotalSales(DateTime initial, DateTime final)
+        public decimal TotalSales(DateTime initial, DateTime final)
         {
             return Sellers.Sum(seller =>  seller.TotalSales(initial, final));
         }
