+++ WebSalesMvcWithAngular/Models/Seller.cs
@@ -59,7 +59,7 @@ namespace WebSalesMvc.Models
             Sales.Remove(sr);
         }
 
-        public double TotalSales(DateTime initial, DateTime final)
+        public decimal TotalSales(DateTime initial, DateTime final)
         {
             return Sales.Where(sr => sr.Date >= initial && sr.Date <= final).Sum(sr => sr.Amount);
         }
